requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pillow>=8.0.0
requests>=2.25.0
openai>=0.27.0
unrealcv>=0.4.0
flask>=2.0.0
pyyaml>=6.0.0
python-dotenv>=0.19.0 
